refactor: implement new InputTagEditor and TagsDisplay components, update tag handling logic and imports
