+++ src/components/Tags/_PostBottomSheet.tsx
@@ -2,10 +2,9 @@
 
 import { PostView } from '@/types/Post';
 import { Input } from '@social/ui-shared';
-import InputTag from './components/_InputTag2';
-import Tags from './components/_Tags2';
+import { Components } from './components';
 import Root from '../ui-shared/lib/BottomSheet/_Root';
-import { useUtilsTag } from './utils/_TagsUtils2';
+import { TagsInteractionUtils } from './utils/_TagsInteractionUtils';
 
 interface TagProps {
   show: boolean;
@@ -17,19 +16,19 @@ interface TagProps {
 }
 
 export function PostBottomSheet({ show, setShow, post, title, className, tagsError }: TagProps) {
-  const { selectedTag } = useUtilsTag(post);
+  const { selectedTag } = TagsInteractionUtils(post);
 
   return (
     <Root show={show} setShow={setShow} title={title ?? 'Tag Post'} className={className}>
       <div className="w-full items-stretch flex-col inline-flex gap-6 -mt-6">
         <div className="flex flex-col md:flex-row gap-6">
-          <InputTag post={post} tagsError={tagsError} />
+          <Components.InputTag post={post} />
           <div
             id="current-tags"
             className="justify-start items-start gap-2 flex flex-col overflow-y-auto min-w-[200px] max-h-[300px] scrollbar-thin scrollbar-webkit"
           >
             <Input.Label value={selectedTag ? 'Tagged' : 'Current tags'} />
-            <Tags post={post} />
+            <Components.TagsDisplay post={post} />
           </div>
         </div>
       </div>
