+++ src/components/Tags/utils/index.tsx
@@ -1,11 +1,11 @@
 import { handleAddTag, handleDeleteTag } from './_TagsCommonFunctions';
-import { useUtilsTag } from './_TagsUtils2';
-import TagsUtils from './_TagsUtils';
+import { TagsInteractionUtils } from './_TagsInteractionUtils';
+import { TagsUtils } from './_TagsUtils';
 
 export const Utils = {
   handleAddTag,
   handleDeleteTag,
-  useUtilsTag,
+  TagsInteractionUtils,
   TagsUtils
 };
